requirements.txt
arxiv>=1.4.8
requests>=2.31.0
pydantic>=2.5.0
PyYAML>=6.0.1
schedule>=1.2.0
python-dateutil>=2.8.2
pdfplumber>=0.10.0
tiktoken>=0.5.1
selenium>=4.15.0
webdriver-manager>=4.0.1
python-dotenv>=1.0.0
tenacity>=8.2.3
markdown>=3.5.1
beautifulsoup4>=4.12.2
lxml>=4.9.3
